Adding pattern data generator.
Pattern data generator is similar to the random generator, but uses a
compressibility factor from 0 (entirely random) to 1.0 (entirely
pattern). This allows for testing of compression at various places in
the web stack.

Note: currently the compressibility is fixed at 0.5 and the data uses
a circular buffer 16MiB in size, so it will repeat after 16MiB.